+++ src/lib/components/common/SplitView.svelte
@@ -70,6 +70,7 @@
 <style>
   .split-view {
     display: flex;
+    width: 100%;
     height: 100%;
     overflow: hidden;
   }